+++ ep1/sequential_multiply.h
@@ -1,1 +1,1 @@
-double sequentialMultiply(double** matrixA, double** matrixB, double** matrixC, int n_rows_a, int n_cols_a, int n_cols_b);
+double sequentialMultiply(double** matrixA, double** matrixB, double** matrixC, long long int n_rows_a, long long int n_cols_a, long long int n_cols_b);
